rill-github-analytics/pyproject.toml
[tool.poetry]
name = "example-github-analytics"
version = "0.1.0"
description = ""
authors = ["Eric P Green <ericpgreen2@gmail.com>"]
readme = "README.md"
packages = [{include = "example_github_analytics"}]

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-storage = "^2.9.0"
python-dotenv = "^1.0.0"
pandas = "^2.0.2"
fsspec = "^2023.6.0"
gcsfs = "^2023.6.0"
pyarrow = "^12.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
